perf(message): React.memo + 발신자 아바타만 좁게 구독
Message 를 React.memo 로 감싸 부모 리렌더/새 메시지 도착 시 props 가 동일한
기존 메시지는 다시 그리지 않게 한다. onlinePeers 배열 전체 구독을 제거하고
발신자의 profileImageUrl(primitive)만 좁게 구독해, 무관한 피어 한 명의 상태
변화에 모든 메시지가 리렌더되던 문제를 없앤다. 반응자 닉네임 툴팁 때문에
onlinePeers/pastDMPeers 가 필요한 리액션 배지는 ReactionBadges 자식 컴포넌트로
분리해, 리액션이 있는 메시지에서만 피어 상태를 구독하도록 범위를 좁힌다.
